Mod was actually not working because event call was in SubModule class instead of CampaignBehavior class. Fixed children growth ratio was not applied properly. Fixed children instant growth. Fixed turkish translation was not applied properly. Added logging for debugging. Added ChildrenGrowFasterCampaignBehavior class. Added unregister from an event after game is closing. Removed SubModuleDebugging class. Changed translation from "Growth Boost By Days" to "Growth Rate" Changed translation from "Adjusts how many days to add to normal growth day. If you set it to 1 it means children will get older twice as fast. Normal day + 1 day = 2 days older in one day. [Default: 2]" to "Children growth rate. 1 = normal rate; 2 = twice as fast; 3 = three times as fast growth rate and so on. [Default: 2]".
