+++ _Module/ModuleData/Languages/CNs/std_module_strings_xml-zho-CN.xml
@@ -8,8 +8,8 @@
 		<string id="CGFR_S7CDl3Ac" text="{randomChild.Name} 获得了特性 {randomTrait.Name}，等级为 {randomTraitLevel}！" />
 		<string id="CGFR_YYayWvdK" text="儿童快速成长 重制版" />
 		<string id="CGFR_5GjemDpY" text="儿童成长速度设置" />
-		<string id="CGFR_xpz8K0H5" text="按天数增加成长速度" />
-		<string id="CGFR_1ZWhgRlk" text="调整添加到正常成长天数的天数。如果设置为 1，意味着儿童的成长速度是正常的两倍。正常天数 + 1 天 = 1 天内长大 2 天。[默认值: 1]" />
+		<string id="CGFR_xpz8K0H5" text="增长率" />
+		<string id="CGFR_1ZWhgRlk" text="儿童生长速度。1 = 正常速度；2 = 两倍；3 = 三倍，以此类推。[默认值：2]" />
 		<string id="CGFR_86Zt3waL" text="英雄成年时的年龄？" />
 		<string id="CGFR_Aog3zurZ" text="让此模组知道儿童成年时的年龄。为什么？因为如果您使用了 'Bannerlord Expanded - Children Expanded' 模组，其中有一个设置 '英雄成年时的年龄'，请将此设置与该设置保持一致。[默认值: 18]" />
 		<string id="CGFR_88NRzszQ" text="仅影响玩家的孩子" />
