+++ _Module/ModuleData/Languages/DE/std_module_strings_xml-ger.xml
@@ -8,8 +8,8 @@
 		<string id="CGFR_S7CDl3Ac" text="{randomChild.Name} hat die Eigenschaft {randomTrait.Name} mit Stufe {randomTraitLevel} erhalten!" />
 		<string id="CGFR_YYayWvdK" text="Kinder wachsen schneller Neuauflage" />
 		<string id="CGFR_5GjemDpY" text="Einstellungen für die Wachstumsrate von Kindern" />
-		<string id="CGFR_xpz8K0H5" text="Wachstumssteigerung nach Tagen" />
-		<string id="CGFR_1ZWhgRlk" text="Legt fest, wie viele Tage zum normalen Wachstumstag hinzugefügt werden. Wenn Sie es auf 1 setzen, bedeutet dies, dass Kinder doppelt so schnell älter werden. Normaler Tag + 1 Tag = 2 Tage älter an einem Tag. [Standard: 1]" />
+		<string id="CGFR_xpz8K0H5" text="Wachstumsrate" />
+		<string id="CGFR_1ZWhgRlk" text="Wachstumsrate bei Kindern. 1 = normale Rate; 2 = doppelt so schnell; 3 = dreimal so schnell usw. [Standard: 2]" />
 		<string id="CGFR_86Zt3waL" text="Wann wird der Held volljährig?" />
 		<string id="CGFR_Aog3zurZ" text="Teilen Sie diesem Mod mit, wann Kinder volljährig werden. Warum? Wenn Sie den Mod 'Bannerlord Expanded - Children Expanded' verwenden, gibt es eine Einstellung 'Held wird volljährig'. Stellen Sie diese Einstellung genauso ein wie diese. [Standard: 18]" />
 		<string id="CGFR_88NRzszQ" text="Nur Kinder des Spielers beeinflussen" />
