+++ _Module/ModuleData/Languages/FR/std_module_strings_xml-fre.xml
@@ -8,8 +8,8 @@
 		<string id="CGFR_S7CDl3Ac" text="{randomChild.Name} a obtenu le trait {randomTrait.Name} avec le niveau {randomTraitLevel} !" />
 		<string id="CGFR_YYayWvdK" text="Children Grow Faster Redux" />
 		<string id="CGFR_5GjemDpY" text="Paramètres du taux de croissance des enfants" />
-		<string id="CGFR_xpz8K0H5" text="Augmentation de la croissance par jours" />
-		<string id="CGFR_1ZWhgRlk" text="Ajuste le nombre de jours à ajouter au jour de croissance normal. Si vous le réglez sur 1, cela signifie que les enfants vieilliront deux fois plus vite. Jour normal + 1 jour = 2 jours plus âgés en un jour. [Par défaut : 1]" />
+		<string id="CGFR_xpz8K0H5" text="Taux de croissance" />
+		<string id="CGFR_1ZWhgRlk" text="Taux de croissance des enfants. 1 = taux normal ; 2 = deux fois plus rapide ; 3 = trois fois plus rapide, et ainsi de suite. [Par défaut : 2]" />
 		<string id="CGFR_86Zt3waL" text="Quand le héros atteint-il l'âge adulte ?" />
 		<string id="CGFR_Aog3zurZ" text="Indiquez à ce mod quand les enfants atteindront l'âge adulte. Pourquoi ? Parce que si vous utilisez le mod 'Bannerlord Expanded - Children Expanded', il y a un paramètre 'Le héros atteint l'âge adulte', faites en sorte que ce paramètre soit identique à celui-ci. [Par défaut : 18]" />
 		<string id="CGFR_88NRzszQ" text="Affecter uniquement les enfants du joueur" />
