+++ _Module/ModuleData/Languages/JP/std_module_strings_xml_jpn.xml
@@ -8,8 +8,8 @@
 		<string id="CGFR_S7CDl3Ac" text="{randomChild.Name} が特性 {randomTrait.Name} をレベル {randomTraitLevel} で獲得しました！" />
 		<string id="CGFR_YYayWvdK" text="子供の成長が早くなる Redux" />
 		<string id="CGFR_5GjemDpY" text="子供の成長率設定" />
-		<string id="CGFR_xpz8K0H5" text="日数による成長ブースト" />
-		<string id="CGFR_1ZWhgRlk" text="通常の成長日に追加する日数を調整します。1に設定すると、子供は通常の2倍の速さで成長します。通常の日 + 1日 = 1日で2日分成長します。[デフォルト: 1]" />
+		<string id="CGFR_xpz8K0H5" text="成長率" />
+		<string id="CGFR_1ZWhgRlk" text="子供の成長率。1 = 通常の速度、2 = 2倍の速度、3 = 3倍の速度など。[デフォルト: 2]" />
 		<string id="CGFR_86Zt3waL" text="英雄が成人するのはいつですか？" />
 		<string id="CGFR_Aog3zurZ" text="このモッドに子供が成人する年齢を知らせてください。なぜですか？『Bannerlord Expanded - Children Expanded』モッドを使用している場合、『英雄が成人する』という設定があります。この設定を同じにしてください。[デフォルト: 18]" />
 		<string id="CGFR_88NRzszQ" text="プレイヤーの子供のみ影響" />
