+++ _Module/ModuleData/Languages/PL/std_module_strings_xml_pol.xml
@@ -8,8 +8,8 @@
 		<string id="CGFR_S7CDl3Ac" text="{randomChild.Name} zdobył cechę {randomTrait.Name} na poziomie {randomTraitLevel}!" />
 		<string id="CGFR_YYayWvdK" text="Dzieci Rosną Szybciej Redux" />
 		<string id="CGFR_5GjemDpY" text="Ustawienia tempa wzrostu dzieci" />
-		<string id="CGFR_xpz8K0H5" text="Przyspieszenie wzrostu według dni" />
-		<string id="CGFR_1ZWhgRlk" text="Dostosowuje liczbę dni dodawanych do normalnego dnia wzrostu. Jeśli ustawisz na 1, oznacza to, że dzieci będą starzeć się dwa razy szybciej. Normalny dzień + 1 dzień = 2 dni starsze w jeden dzień. [Domyślnie: 1]" />
+		<string id="CGFR_xpz8K0H5" text="Tempo wzrostu" />
+		<string id="CGFR_1ZWhgRlk" text="Tempo wzrostu dzieci. 1 = normalne tempo; 2 = dwukrotnie szybsze; 3 = trzykrotnie szybsze tempo wzrostu itd. [Domyślnie: 2]" />
 		<string id="CGFR_86Zt3waL" text="Kiedy bohater osiąga pełnoletność?" />
 		<string id="CGFR_Aog3zurZ" text="Poinformuj ten mod, kiedy dzieci osiągną pełnoletność. Dlaczego? Ponieważ jeśli używasz moda 'Bannerlord Expanded - Children Expanded', istnieje ustawienie 'Bohater osiąga pełnoletność', ustaw je tak samo jak tutaj. [Domyślnie: 18]" />
 		<string id="CGFR_88NRzszQ" text="Wpływ tylko na dzieci gracza" />
