+++ _Module/ModuleData/Languages/RU/std_module_strings_xml_rus.xml
@@ -8,8 +8,8 @@
 		<string id="CGFR_S7CDl3Ac" text="{randomChild.Name} получил черту {randomTrait.Name} уровня {randomTraitLevel}!" />
 		<string id="CGFR_YYayWvdK" text="Дети растут быстрее Ремастер" />
 		<string id="CGFR_5GjemDpY" text="Настройки скорости роста детей" />
-		<string id="CGFR_xpz8K0H5" text="Ускорение роста по дням" />
-		<string id="CGFR_1ZWhgRlk" text="Настраивает, сколько дней добавлять к обычному дню роста. Если установить 1, это означает, что дети будут стареть в два раза быстрее. Обычный день + 1 день = 2 дня старше за один день. [По умолчанию: 1]" />
+		<string id="CGFR_xpz8K0H5" text="Темпы роста" />
+		<string id="CGFR_1ZWhgRlk" text="Скорость роста детей. 1 = нормальная скорость; 2 = в два раза быстрее; 3 = в три раза быстрее скорость роста и т. д. [По умолчанию: 2]" />
 		<string id="CGFR_86Zt3waL" text="Когда герой становится взрослым?" />
 		<string id="CGFR_Aog3zurZ" text="Сообщите этому моду, когда дети станут взрослыми. Почему? Потому что если вы используете мод 'Bannerlord Expanded - Children Expanded', там есть настройка 'Герой становится взрослым', сделайте эту настройку такой же, как и здесь. [По умолчанию: 18]" />
 		<string id="CGFR_88NRzszQ" text="Влиять только на детей игрока" />
