+++ _Module/ModuleData/Languages/SP/std_module_strings_xml-spa-M9.xml
@@ -8,8 +8,8 @@
 		<string id="CGFR_S7CDl3Ac" text="¡{randomChild.Name} ha ganado el rasgo {randomTrait.Name} con nivel {randomTraitLevel}!" />
 		<string id="CGFR_YYayWvdK" text="Los Niños Crecen Más Rápido Redux" />
 		<string id="CGFR_5GjemDpY" text="Configuración de la Tasa de Crecimiento de los Niños" />
-		<string id="CGFR_xpz8K0H5" text="Impulso de Crecimiento por Días" />
-		<string id="CGFR_1ZWhgRlk" text="Ajusta cuántos días agregar al día de crecimiento normal. Si lo configuras en 1, significa que los niños envejecerán el doble de rápido. Día normal + 1 día = 2 días más viejos en un día. [Por defecto: 1]" />
+		<string id="CGFR_xpz8K0H5" text="Índice de crecimiento" />
+		<string id="CGFR_1ZWhgRlk" text="Tasa de crecimiento infantil. 1 = tasa normal; 2 = el doble de rápido; 3 = el triple de rápido, y así sucesivamente. [Predeterminado: 2]" />
 		<string id="CGFR_86Zt3waL" text="¿Cuándo el Héroe Alcanza la Mayoría de Edad?" />
 		<string id="CGFR_Aog3zurZ" text="Haz saber a este mod cuándo los niños alcanzarán la mayoría de edad. ¿Por qué? Porque si usas el mod 'Bannerlord Expanded - Children Expanded', hay una configuración 'El Héroe Alcanza la Mayoría de Edad', haz que esta configuración sea igual a esa. [Por defecto: 18]" />
 		<string id="CGFR_88NRzszQ" text="Afectar Solo a los Hijos del Jugador" />
